upd: create markdown

+++ index.js
@@ -1,26 +1,44 @@
-const cheerio = require("cheerio");
-const fs = require("fs");
-const axios = require("axios");
-const express = require("express");
-var TurndownService = require("turndown");
-var markdown = require("markdown-js");
+const cheerio = require('cheerio');
+const fs = require('fs');
+const axios = require('axios');
+const express = require('express');
+const TurndownService = require('turndown');
+const markdown = require('markdown-js');
 const app = express();
-var turndownService = new TurndownService();
+const config = require('./config.json');
+const turndownService = new TurndownService({
+  headingStyle: 'atx',
+  bulletListMarker: '-',
+});
+
+turndownService.addRule('PreCode', {
+  filter: ['pre'],
+  replacement: function(content, node) {
+    return '\n```\n' + content + '\n```\n';
+  },
+});
 
 function getData(time, language) {
-  let url = "https://github.com/yanyue404/blog/issues/1"; // 拼接请求的页面链接
+  var fileDirectory = 'test/';
+  let url = config.github.blog + '/issues/1'; // 拼接请求的页面链接
   return axios
     .get(url)
     .then(function(response) {
       let html_string = response.data.toString(); // 获取网页内容
       const $ = cheerio.load(html_string); // 传入页面内容
       // console.log($("table").html());
-      var content = turndownService.turndown($("table").html());
+      var content = turndownService.turndown($('table').html());
       console.log(content);
-      fs.writeFile("./test/git入门与实践.md", content, err => {
-        if (err) throw err;
-        console.log("It's saved!");
-      });
+      // 判断文件夹路径是否存在
+      if (fs.existsSync(fileDirectory)) {
+        fs.writeFile('./test/git入门与实践.md', content, err => {
+          if (err) throw err;
+          console.log("It's saved!");
+        });
+      } else {
+        console.log(fileDirectory + '  Not Found!');
+      }
+
       return Promise.resolve(content);
     })
     .catch(function(error) {
@@ -28,8 +46,8 @@ function getData(time, language) {
     });
 }
 
-app.get("/", (req, res) => {
-  let promise = getData("daily"); // 发起抓取
+app.get('/', (req, res) => {
+  let promise = getData('daily'); // 发起抓取
   promise.then(response => {
     //markdown.markHtml(); 是将markdown格式的字符转换成Html
     var html = markdown.makeHtml(response);
@@ -38,4 +56,4 @@ app.get("/", (req, res) => {
   });
 });
 
-app.listen(3000, () => console.log("Listening on port 3000!")); // 监听3000端口
+app.listen(3000, () => console.log('Listening on http://localhost:3000!')); // 监听3000端口
